Changed Cache to global array of caches, to enable global checking of state.
CHanged the macros for cache manipulation so they are easier to use, most of the work is now done inside the macro.

Added cache state cheking process, which should check the validity of the cache state
between every step.

Will have to change the structure of the main loop, as the bus operations CANNOT be
inside expressions, as they have side effects.