Finished implementation, verified all that I can think of.
Changed cache back to global, so that it can be verified by LTL.
Accessing locals is does not work well with partial order reduction.

Moved execute_in_cache to be executed whenever it can, before snooping
on the bus. Before, it could decide whether to snoop or to execute,
which lead to non-progression cycle of always snooping.